feat: show all outlet for super user

+++ src/pages/api/admin/category/[id]/index.js
@@ -1,6 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 
 import { getSession } from '@auth0/nextjs-auth0'
+import buildQuery from 'utils/buildQuery'
 import { supabase } from 'utils/supabaseClient'
 
 export default async function handler(req, res) {
@@ -8,11 +9,11 @@ export default async function handler(req, res) {
 
   // get single outlet data
   if (req.method == 'GET') {
-    let { data: category, error } = await supabase
-      .from('category')
-      .select('*')
-      .eq('id', req.query.id)
-      .eq('user_id', user.sub)
+    let { data: category, error } = buildQuery(
+      user,
+      await supabase.from('category').select('*').eq('id', req.query.id),
+      await supabase.from('category').select('*').eq('id', req.query.id).eq('user_id', user.sub)
+    )
 
     if (category) {
       res.status(200).json(category[0])
@@ -23,11 +24,18 @@ export default async function handler(req, res) {
 
   // update single outlet data
   if (req.method == 'PUT') {
-    const { data } = await supabase
-      .from('category')
-      .update({ ...req.body, updated_at: new Date() })
-      .eq('id', req.query.id)
-      .eq('user_id', user.sub)
+    const { data } = buildQuery(
+      user,
+      await supabase
+        .from('category')
+        .update({ ...req.body, updated_at: new Date() })
+        .eq('id', req.query.id),
+      await supabase
+        .from('category')
+        .update({ ...req.body, updated_at: new Date() })
+        .eq('id', req.query.id)
+        .eq('user_id', user.sub)
+    )
 
     if (data) {
       res.status(200).json({ ok: true, message: 'Data berhasil diupdate' })
@@ -38,13 +46,21 @@ export default async function handler(req, res) {
 
   // delete single data
   if (req.method == 'DELETE') {
-    await supabase.from('product').delete().eq('category_id', req.query.id).eq('user_id', user.sub)
+    buildQuery(
+      user,
+      await supabase.from('product').delete().eq('category_id', req.query.id),
+      await supabase
+        .from('product')
+        .delete()
+        .eq('category_id', req.query.id)
+        .eq('user_id', user.sub)
+    )
 
-    const { data: category } = await supabase
-      .from('category')
-      .delete()
-      .eq('id', req.query.id)
-      .eq('user_id', user.sub)
+    const { data: category } = buildQuery(
+      user,
+      await supabase.from('category').delete().eq('id', req.query.id),
+      await supabase.from('category').delete().eq('id', req.query.id).eq('user_id', user.sub)
+    )
 
     if (category) {
       res.status(200).json({ ok: true, message: 'Data berhasil dihapus' })
